tests/wpt/tests/html/canvas/offscreen/text/2d.text.direction.valid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.text.direction.valid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.direction.valid</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.direction = 'ltr';
  _assertSame(ctx.direction, 'ltr', "ctx.direction", "'ltr'");

  ctx.direction = 'rtl';
  _assertSame(ctx.direction, 'rtl', "ctx.direction", "'rtl'");

  ctx.direction = 'inherit';
  _assertSame(ctx.direction, 'inherit', "ctx.direction", "'inherit'");
  t.done();

});
</script>
